add switch url to IP if using local connection

+++ lib/screens/taking_order_vendor/payment/payment_main_page.dart
@@ -49,8 +49,8 @@ class PaymentMainPage extends StatelessWidget {
                           ),
                         ) :
                         PiutangCard(nmtoko : _takingOrderVendorController.nmtoko.value,
-                        totalpiutang: _takingOrderVendorController.totalpiutang.value == "0" || _takingOrderVendorController.totalpiutang.value == "null" ? "0" : Utils().formatNumber(int.parse( _takingOrderVendorController.totalpiutang.value)),
-                        totaljatuhtempo: _takingOrderVendorController.totaljatuhtempo.value == "0" || _takingOrderVendorController.totaljatuhtempo.value == "null" ? "0" : Utils().formatNumber(int.parse( _takingOrderVendorController.totaljatuhtempo.value)))),
+                        totalpiutang: _takingOrderVendorController.totalpiutang.value == "0" || _takingOrderVendorController.totalpiutang.value == "null" ? "0" : Utils().formatNumber(double.parse( _takingOrderVendorController.totalpiutang.value)),
+                        totaljatuhtempo: _takingOrderVendorController.totaljatuhtempo.value == "0" || _takingOrderVendorController.totaljatuhtempo.value == "null" ? "0" : Utils().formatNumber(double.parse( _takingOrderVendorController.totaljatuhtempo.value)))),
                     const SizedBox(
                       height: 15,
                     ),
